Moved ChangedItem to its own file
